refactor(inspectorWidgets): rename filenames of inspectorWidget components using PascalCase (boolWidget.jsx -> BoolWidget.jsx)
