+++ app/components/inspectorWidgets/BoolWidget.jsx
