+++ app/components/inspectorWidgets/HintWidget.jsx
