+++ app/components/inspectorWidgets/NumberWidget.jsx
