+++ app/components/inspectorWidgets/StringWidget.jsx
